Update Package.swift and podspecs with new checksums and sha1s for v12.0.1
Summary:
I ran this on the release branch: remote/fbobjc/releases/release-ios-sdk-v12.0.1
```
swift run runner prepare-release | xcpretty
```

This diff was created on the release branch but I'm going to rebase it on master so that it can be merged into master. (It needs to be in master in case someone points their SPM to master, that way they still get the latest release).

After its merged into master, I will pick it into the release branch.

Reviewed By: samodom

Differential Revision: D31696499

fbshipit-source-id: 2a0880489e10fb8457edbfd021be227a546ee859