Make `init` and `new` unavailable for SuggestedEventsIndexer
Summary:
In ObjC it's easy to provide convenience initializers that do not call a designated initializer. This is a step towards making it more difficult to bypass required initializers by making `new` and `init` unavailable when there are convenience init or factory methods / singletons available.

This forced a change to the tests since we can no longer create a new instance directly but must do it through the singleton.

Reviewed By: samodom

Differential Revision: D31735827

fbshipit-source-id: 942b04308ae47c687dd89a69638f2ab8037dd1fb