Make `init` and `new` unavailable for FeatureManager
Summary:
In ObjC it's easy to provide convenience initializers that do not call a designated initializer. This is a step towards making it more difficult to bypass required initializers by making `new` and `init` unavailable when there are convenience init or factory methods / singletons available.

This forced a change to the tests since we can no longer create a new instance directly but must do it through the singleton.

Reviewed By: samodom

Differential Revision: D31735826

fbshipit-source-id: fbb5c29ed4387eb0304d9e0ba03eff16d4dbcc15