Fix archiving on Xcode 12.5
Summary:
It's weird but `@_exported import` does not seem to work when archiving with SPM on Xcode 12.5. This means that symbols are not re-exported under the library names that SPM defines but instead are only available in the underlying `FBSDK`-prefixed modules.

## Some context:
The SPM targets with the `Facebook` prefix (ex: `FacebookCore` are by and large just hacks to help people avoid importing dependent frameworks and libraries such as `Accelerate` and `z`. Essentially they are just wrappers. We were using `@_exported import` so that you could use the symbols from the 'wrapped' framework as though they were coming from the original framework.

This works in Xcode13 but does not work for archiving on 12.5.

If you change the import to be `FBSDK` prefixed, you can still reference all of the symbols you need when using 12.5.

Additionally this changeset fixes the need to explicitly include `FacebookAEM` and `FacebookBasics` in your project.

Reviewed By: jawwad

Differential Revision: D31840198

fbshipit-source-id: 145a904c1408cb8fda3123db185d161485c4607c