Main (#1923)
Summary:
Thanks for proposing a pull request!

To help us review the request, please complete the following:

- [x ] sign [contributor license agreement](https://code.facebook.com/cla)
- [ x] I've ensured that all existing tests pass and added tests (when/where necessary)
- [ x] I've updated the documentation (when/where necessary) and [Changelog](CHANGELOG.md) (when/where necessary)
- [x ] I've added the proper label to this pull request (e.g. `bug` for bug fixes)

## Pull Request Details

Enable 'needs-attention' github action. This action uses the labels 'needs-triage' and 'needs-feedback' to indicate the state of the issue. 'needs-feedback' is used when more information is required to start the resolution. 'needs-triage' indicates that a repo administrator has to triage the issue. This pull request also prevents stale action to close issues tagged with 'needs-triage' and 'acknowledged' labels.

Pull Request resolved: https://github.com/facebook/facebook-ios-sdk/pull/1923

Reviewed By: jawwad

Differential Revision: D31898810

Pulled By: xilosada

fbshipit-source-id: 4fcc2f098593fbdded63fd6e946c0e16dd9c1e97