Removing Custom Score And Adding Tournament To Show
Summary:
After talking to some of the team about including score in tournaments we decided its not a good idea. We also decided to not support a custom score yet, so all scores will be int values.

In this Diff, I'm removing  the score parameter which use to be any object that conforms to the score protocol. I'm replacing it with an integer type, which should be a timestamp or integer score. I also move the tournament parameter from the init to the show method which will be the method for showing the share dialog for updating an existing tournament.

The reason for this move is in a future diff I'll add another method where we can pass a score and tournamentConfig to show the share dialog to share a newly created tournament with the config. This means a share dialog doesn't use a tournament object to create a new tournament.

Reviewed By: ainsleyh

Differential Revision: D31831695

fbshipit-source-id: f463059c7d9588164a5096f5c40c8a15771b21fc