Removes Score Protocol and Any Conforming Types
Summary: Tournament will not hold the score value for the user and we aren't accepting custom score types as discussed in earlier diffs, with ainsleyh and the rest of the tournaments product team. This means there is no longer a use for the score protocol and any types that conform to it.

Reviewed By: ainsleyh

Differential Revision: D31840237

fbshipit-source-id: c363e59f23c4c25a8b97eb05f9e519d2170aba24