Adds a Error/Completion Handling for ShareDialog
Summary:
This adds the logic that parses the returned deeplink from the fb/msite `ShareTournamentDialog`. If the url contains a tournament_ID it means that the user successfully completed the `ShareTournamentDialog` flow. We will create a tournament based using the returned ID and call dialogDidComplete with the created tournament.

If we are updating a tournament we check the returned ID matches the passed in tournament ID and return the same tournament with dialogDidComplete.

This functionally is tested using hackbook.

Reviewed By: ainsleyh

Differential Revision: D31874422

fbshipit-source-id: 66b0a988c087ba572f2867875a5928cda018168c