AppLinkUtility.fetchDeferredAppLink never calling completion on error (#1903)
Summary:
Thanks for proposing a pull request!

To help us review the request, please complete the following:

- [x] sign [contributor license agreement](https://code.facebook.com/cla)
- [x] I've ensured that all existing tests pass and added tests (when/where necessary)
- [x] I've updated the documentation (when/where necessary) and [Changelog](CHANGELOG.md) (when/where necessary)
- [] I've added the proper label to this pull request (e.g. `bug` for bug fixes)

## Pull Request Details

Completion handles are called now even when an error occurs. Due to issue where the following code has to be executed on the main thread. When Facebook down as it was last night the SDK stops working and as a result the main thread was blocked. This ensures that the completion is still called regardless.

`AppLinkUtility.fetchDeferredAppLink { url, _ in
            completion(url)
}`

Pull Request resolved: https://github.com/facebook/facebook-ios-sdk/pull/1903

Test Plan:
* Updated two unit tests regarding these completions
* Tested manually by launching application with the above last night when Facebook was down and worked as expected. This can also be tested if you launch the application with above code with aeroplane mode turned on.

Reviewed By: samodom

Differential Revision: D31743390

Pulled By: joesus

fbshipit-source-id: 2b5eb79eb723004fe08de6297c16985bf11f364d