Removes Context Choose url builder for msite
Summary: Before we were using two controllers for the context choose dialog on iOS, one for the msite and one for app switching. I have a backend diffs that reuses a the app-switching controller for msite and gets rid of the msite controller. This diffs cleans up the sdk code that creates the msite controller URL.

Reviewed By: samodom

Differential Revision: D31507043

fbshipit-source-id: e1bf0f1509f0bfe9e1e3eca7098a0913529d4223