Convert the FBSDKFileHandling protocol to Swift
Summary:
This is an internal type that needs to be made public for the transition to Swift, otherwise it won't be visible to Objective-C code that uses it.
Named the Swift type with an initial underscore, i.e. _FileHandling to signify that it is internal. Also added a comment noting that it is an internal class and shouldn't be used.

Reviewed By: joesus

Differential Revision: D32328419

fbshipit-source-id: fe50dc67b7d315db10e5d2abd60094a0b49435f9