Update AppEventsDeviceInfo callsites
Summary:
The `extendDictionaryWithDeviceInfo` took a dictionary and added the encoded device info to it under a known key.
This gets rid of that convenience method and instead allows callers to construct their own device info object and get the encoded events and key from it. This feels a little more flexible.

Additionally it conforms it to a new protocol `DeviceInformationProviding` so that it can be injected into the two types that use it.

Reviewed By: samodom

Differential Revision: D32357944

fbshipit-source-id: a9a62ddf6954b6f9f582fad9ee98939cd523e6f0