Convert LoginManagerTests to Swift
Summary:
Converts the tests that makes sense to convert. Some of them were testing situation that cannot be reproduced using Swift.

Actual production code change:

* Fixed nullability where it was clearly incorrect
* redefined some getter methods as read only properties
* added some swift names where it was very unclear what the method meaning was

Mostly tried to keep the changes to the tests.

One weird test deletion that might raise questions. Deleted: `testOpenURLReauthNoPermissionsIsNotCancelled`
This test created a token with "public_profile" and "read_stream" permissions. Then called `validateReauthentication` with that token which will always create a graph request. Then it checked that a graph request was created.
It then called open url but since there was no handler set up, there was nothing to check...
If this was testing a valid use case I could not figure what it was from the test.

Reviewed By: samodom

Differential Revision: D32581692

fbshipit-source-id: 12d1f5a683256a343c18edb87b9a0329371547bd