Convert FBSDKCameraEffectArgumentsTests to Swift
Summary: The XCTestCase+Assertions.h/m files are similar to the XCTestCase+Exceptions.h/m files in FBSDKCoreKit. Not sure of a way to share these. These can't be added to TestTools since it isn't a test target and so dosen't link XCTest.framework, and XCTest.framework can't be linked to manually either.

Reviewed By: samodom

Differential Revision: D32632268

fbshipit-source-id: a6e14430da3d942c00e4045eb96ce3cb09546cfd