Updates Gaming payload and the payload observer
Summary:
This diff cleans up the payload parsing logic. It makes it easy to add new keys we need to listen for when parsing a gaming entry url into a native game.

To add a new listener:
Add it to the keys list.
Create a new delegate Method
Check for the key in the GameEntryInfo and The third party app implemented the delegate method for said key
If so call the delegate method.

Reviewed By: jawwad

Differential Revision: D32808171

fbshipit-source-id: 0af111cc7e728f386a0def06d7a43421cab121f7