Fix swiftformat extra-indenting chained statements
Summary:
SwiftFormat is indenting forEach and other chained calls to match the indentation of Xcode 13. In Xcode 12 these weren't indented this way.

It indents the complete block which feels weird. We have the "--xcodeindentation enabled" option in our .swiftformat file and we can disable that, however doing this causes other issues.

One workaround is to move the .forEach call up one line, however the team has decided that this option makes it more difficult to read so we'll go with the default Xcode indentation.

Reviewed By: samodom

Differential Revision: D32852487

fbshipit-source-id: dbca9da47a90d7d9fc3c3dd6e1d47526e72e620d