Update Package.swift and podspecs with new checksums and sha1s for v12.2.1
Summary:
I ran this on the release branch: remote/fbobjc/releases/release-ios-sdk-v12.2.1
```
time swift run --build-path=`mkscratch path` runner prepare-release | xcpretty
```

I created this diff on the release branch but I'm rebasing this onto master first. It needs to be in master in case someone points their SPM to master, that way they still get the latest release. Once it's in master it will be picked into the release branch.

Reviewed By: samodom

Differential Revision: D32988695

fbshipit-source-id: 12fa9804177e9ab391b97f69b5460ae1805d6689