SwiftLint: Disable the opt-in (and quirky) indentation_width rule
Summary:
SwiftLint's indentation_width rule is an opt-in rule that works great 99% of the time but it has it's quirks for which we've had to use some workarounds, for example having the "guard" keyword in a multiline guard on its own line. At other times we've just included the "// swiftlint:disable" statemets to selectively disable the rule.

Now that we are using SwiftFormat to automatically format swift code, we no longer need to have the indentation_width rule enabled in .swiftlint.yml

Reviewed By: samodom

Differential Revision: D33035634

fbshipit-source-id: 534a46d447b2ce0290b1e858ed5405b62c57e35e