Make the xcodegen --use-cache flag be opt-in by setting the XCODEGEN_USE_CACHE environment variable
Summary:
The `--use-cache` flag for xcodegen is great but it has some nuances that you might need to know about. For example, if only the project file has changed (say by re-ordering the position of a file in the project within Xcode) then projects won't be re-generated.

This is an issue because changing anything manually in the project file causes it to become invalid. And running ./generate-projects.sh will not regenerate the project if files haven't been added or removed. (Usually, I get something like: `Undefined symbol: _OBJC_CLASS_$_FBSDKBasicUtility`)

{F686793297}

This changes the flag to be opt-in. You can set the `XCODEGEN_USE_CACHE` variable to anything and it will use the `--use-cache` flag.

How to use the flag
```
# 1. Set it everytime you want to run the script
XCODEGEN_USE_CACHE=1 ./generate-projects.sh

# 2. Set it once per session
export XCODEGEN_USE_CACHE=1
./generate-projects.sh # uses cache
./generate-projects.sh # uses cache

# 3. Set it in your .bashrc or .zshrc so that it's always set.
```

If you run into any issues you can reset the cache with the following command:
```
rm -rf ~/.xcodegen/cache
```

Reviewed By: joesus

Differential Revision: D33106854

fbshipit-source-id: 1dcd477b0b933e9f6efb4e093ac84f50e11f3704