Simplify our swiftlint config due to nested configurations not working recursively
Summary:
We've decided that we don't really care about the type_body_length, file_length, and function_body_length rules and in reality we would just disable them instead of shortening the type, file or function.

These were initially set to 10x default values only for tests but I'm getting rid of the test configs and moving this into the main config.

Reviewed By: joesus

Differential Revision: D33113747

fbshipit-source-id: 381ba70ae076fe9897a35318fb989f3312ba5995