Fix for sharing canceled when Facebook app not installed. Fixes #1976
Summary:
Fix for GitHub Issue: https://github.com/facebook/facebook-ios-sdk/issues/1976

Sharing was canceled when the device didn't have a Facebook app installed.

Encountered an issue where an uninitialized Bool in the sharing flow was being set to true (the value in the debugger and log statement was `13`) when optimizations were enabled in the release configuration.

In the debug configuration it was uninitialized and thus defaulted to 0.

This explicitly sets uninitialized Bools to false to try and avoid future issues.

Reviewed By: jawwad

Differential Revision: D33130071

fbshipit-source-id: 40121bb1e217d36931d8fe8e409d7a6b6027642b