Adding ObjC linker flag for regression tests
Summary:
Adds the ObjC linker option. Per [Apple Docs](https://developer.apple.com/library/archive/qa/qa1490/_index.html):

> Passing the -ObjC option to the linker causes it to load all members of static libraries that implement any Objective-C class or category. This will pickup any category method implementations. But it can make the resulting executable larger, and may pickup unnecessary objects. For this reason it is not on by default.

We need to make sure that the SDK still works when this aggressive style of linking is on so we're adding the ObjC flag. This should expose any issues we have with linking.

I'm not entirely clear on the issues this could obscure but it feels like overkill to have a duplicate job just for this one flag.

Reviewed By: samodom

Differential Revision: D33255584

fbshipit-source-id: 49ac804518219ccec1b736959887c83efe50c781