Fix issue with _extractOffsetsFromResultDictionary in tests not returing nil when the method actually does
Summary:
The actual _extractOffsetsFromResultDictionary in FBSDKVideoUploader.m returns a nullable dict. However in FBSDKVideoUploader+Testing the interface is defined to return a non-nullable value.

Due to this, when the actual method returns nil, somehow the tests get an empty dictionary.

Fixed the method signature and the tests to check against what is actually returned.

Reviewed By: samodom

Differential Revision: D33699770

fbshipit-source-id: 0ea72cbc6de68da140efaee47f79c17b178a0573