Fix issue in which build_phase_swiftlint.sh would try to lint a deleted file
Summary:
The issue is that it was trying to lint a file that was added in a previous (unlanded) commit and then was removed in the working copy resulting in an error:

```
error: File not found at /Users/jawwad/fbsource/fbobjc/ios-sdk/FBSDKShareKit/FBSDKShareKit/GameRequestActionType.swift.
```

Here is how the script works

It first gets a list of previously committed files not already in master (i.e. unlanded)
```
hg status --modified --added --no-status --rev 'ancestor(master,.)::.'
```

In this case GameRequestActionType.swift was added in a previous commit.

It then tries to get a list of files that have uncommitted changes in the working copy with the following command:

```
hg status --no-status --unknown --modified --added
```

Note that the command doesn't include the `--removed` flag.

The script then combines the 2 lists and tries to lint them, however, the previouly-added-but-now-deleted file causes an error.

Note: If the working copy was committed with the the error would not occur since the deleted file wouldn't show up in the previously committed (but unlanded) changes.

Fixing the issue by checking to see if a file exists before linting it.

Reviewed By: samodom

Differential Revision: D33796302

fbshipit-source-id: fc1bbffb7b038ce1a2d9ffb8e761f427e683b53f