Fix check_for_xcodebuild_warnings.py not working due to hardcoded OS of 15.0
Summary:
The destination specifier was hardcoded to use 15.0. However now that 15.0 is no longer available it stopped working. Removed the hardcoded OS so that it would just use what is available.

The error was the following:
```
Failed to run xcodebuild. Return code: 70
xcodebuild: error: Unable to find a destination matching the provided destination specifier:
		{ platform:iOS Simulator, OS:15.0, name:iPhone 13 }

	The requested device could not be found because no available devices matched the request.

	Available destinations for the "BuildAllKits-Dynamic" scheme:
        ...
        ...
        ...
		{ platform:iOS Simulator, id:F3A225D9 (https://github.com/facebook/facebook-ios-sdk/commit/266c98570d6ada84404253127184a35feb2f878b)-91C5-459E-B599-59E19F22C85E, OS:15.2, name:iPhone 13 }
        ...
        ...
        ...
```

Reviewed By: samodom

Differential Revision: D33897972

fbshipit-source-id: 26c803e7d96fdc3938a9d009d2ea81c269e204b1