Complete restructuring/renaming of FacebookGamingServices to FBSDKGamingServicesKit
Summary:
Context:
We currently have 2 gaming modules FacebookGamingServices and FBSDKGamingServicesKit

FacebookGamingServices is currently the main module and FBSDKGamingServicesKit is a wrapper around it. The FBSDKGamingServicesKit wrapper previously included 2 Objective-C wrappers to expose functionality in Swift-only code in FacebookGamingServices. However, the maintenance cost of maintaining 2 separate modules was very high so these are being combined into one module.

For this, all of the Objective-C wrappers have been moved to FacebookGamingServices so FBSDKGamingServicesKit is now just an empty wrapper around FacebookGamingServices.

In this diff, I've made the FBSDKGamingServicesKit module the main module and made FacebookGamingServices be a wrapper around it, as we are doing for other kits for Swift Package Manager.

Reviewed By: samodom

Differential Revision: D33935669

fbshipit-source-id: bec45c4534d30800f7150a94a7556d75f78e67a8