Rename SharingValidation protocol to SharingValidatable
Summary:
This protocol requires a single validation method for sharable content and media, but its name does not conform to best practices.

The risk of this being a breaking change is low, but to be on the safe side there is a new, deprecated type alias from the old name to the new one.

This change also makes it easier to abstract validation methods defined in _ShareUtility into a ShareValidating protocol with less ambiguity about how the two protocols are used.

Reviewed By: jawwad

Differential Revision: D34555102

fbshipit-source-id: 7900a082722fdd655513f2f85e65fc0248c29044