Allow project file generation to proceed with a warning even if EXPECTED_XCODEGEN_VERSION does not not match the installed version.
Summary:
In general there should be no issues with generating the project file with a "close enough" version of Xcodegen

Also homebrew makes it difficult to install previous versions of a formula. Previously you could install an older version of a formula by using the path to the raw file on GitHub but that doesn't seem to work anymore. So if someone has a newer version of xcodegen it will be difficult for them to install an older version in order to compile the project files.

This chanege allows the project file generation to proceed and will just display a warning. Example warning:
```
WARNING: Expected xcodegen version is 2.24.0. You have 2.25.0. Project file generation will proceed but you may encounter unexpected issues.
```

Reviewed By: linmx0130

Differential Revision: D34728086

fbshipit-source-id: 568879f4fd6cb2a0cab4701e0562abe56e9df1fb