Support deeplinking using universal links
Summary:
## See the following [Proposal](https://fburl.com/gdoc/z90ch5in)

## Problem

We want to add support for universal links.
In order to do that, developers who have app events integration (AEM) will be required to add a call to the `openURL:` on the SDK delegate class from `application(_:continue:restorationHandler:)` in the short term.

We would like to simplify things for app developers since extracting the webpage URL from NSUserActivity may cause confusion and errors for deep linking.

## Solution

The better way is to just introduce a method with a similar signature to `application(_:continue:restorationHandler:)` in the SDK delegate.
Developers could simply invoke that method in order to have the universal link processed.

By doing that, we will to make FB SDK more accessible to developers.

## Testing

See the test plan section

Reviewed By: samodom

Differential Revision: D34786435

fbshipit-source-id: 289fea5ecc9e106fe0c9fcc94bf2a65b0c07c0ab