Temporarily make Objective-C files in FBAEMKit public in preparation for Swift conversion
Summary:
Currently most files in FBAEMKit are internal. For Objective-C and Swift interoperability within a framework, the files need to be public. If a converted Swift file is not public it will not be visible from Objective-C. Also if Objective-C files are not publicly exposed they will not be visible to Swift.

This commit makes all Objective-C files in FBAEMKit public so that they can be converted to Swift. Made the following changes.

- Added an underscore to the Swift version of the name. Did not add it to the Objective-C version to avoid an extra rename and because these will be converted quickly as they are smaller files.
- Added the "warning INTERNAL - DO NOT USE" warnings
- Quoted imports needed to be updated: #import "FBAEMRule.h" -> #import <FBAEMKit/FBAEMRule.h>

Reviewed By: samodom

Differential Revision: D35150934

fbshipit-source-id: 5b87ad2bb7aa8ef4c3473d807f0592df68acb86e