Refactor testing of NSCoding in AccessTokenTests
Summary:
This is a proof of concept on an updated approach to test that encoding and decoding of NSCoding supporting objects happens correctly.

**TLDR**: Instead of testing NSCoder encoding and decoding with a subclass of NSCoder, this tests encoding and decoding by encoding an object into data and then decoding that data back into an object and making sure that all of the properties are equal.

**Background**:
The impetus for this change was triggered by the issue that the `FBSDKTestCoder` class cannot be fully converted to Swift. There are 2 methods in the class that are not possible to convert. See the following methods and the errors they show

```
final class TestCoder: NSCoder {
  override func decodeObject<DecodedObjectType>(
    of cls: DecodedObjectType.Type,
    forKey key: String
  ) -> DecodedObjectType? where DecodedObjectType: NSObject, DecodedObjectType: NSCoding {
    // Error1: instance method 'decodeObject(of:forKey:)' is declared in extension of 'NSCoder' and cannot be overridden
    // Error2: Overriding non-open instance method outside of its defining module
  }

  override func decodeObject(
    of classes: [AnyClass]?,
    forKey key: String
  ) -> Any? {
    // Error1: Non-@objc instance method 'decodeObject(of:forKey:)' is declared in extension of 'NSCoder' and cannot be overridden
    // Error2: Overriding non-open instance method outside of its defining module
  }
}
```

The `FBSDKTestCoder` class tests both encoding and decoding. For encoding it overrides various methods of `NSCoder` for example like the following:
```
- (void)encodeObject:(id)object forKey:(NSString *)key
{
  self.encodedObject[key] = object;
}
```

Then it will have a test that checks if the `encodedObject` dictionary has that property set under the key that it is encoded with:
```
    XCTAssertEqual(
      coder.encodedObject["tokenString"] as? String,
      SampleAccessTokens.validToken.tokenString,
      "Should encode the expected tokenString with the correct key"
    )
```

For testing decoding its a bit weirder. It saves the class of the decoded object and not the object itself:

```
- (nullable id)decodeObjectOfClass:(Class)aClass forKey:(NSString *)key
{
  self.decodedObject[key] = aClass;

  return nil;
}
```

And in some cases it just saves the method that was called
```
- (nullable id)decodeObjectForKey:(NSString *)key
{
  self.decodedObject[key] = @"decodeObjectForKey";

  return nil;
}
```

Then there are tests for these that check the type that was saved:

```
    XCTAssertTrue(
      decoder.decodedObject["tokenString"] is NSString.Type,
      "Should decode the expected type for the tokenString key"
    )
    XCTAssertTrue(
      decoder.decodedObject["permissions"] is NSSet.Type,
      "Should decode the expected type for the permissions key"
    )
```

It feels a bit weird to check the decoding by saving the type of object that is to be decoded. This means that the object doesn't have to be correct, just the type. So if 2 NSString objects were interchanged and saved with incorrect keys it wouldn't matter.

Reviewed By: samodom

Differential Revision: D35265036

fbshipit-source-id: b07543ccfef6b0961fd7277bcdd79f100f737a8f