Eliminate use of FBSDKTestCoder from FBAEMKit/FBAEMKitTests/AEMInvocationTests.swift
Summary:
This diff also ads a TestUtils.encodeAndDecode(_:) method to reduce some boilerplate in tests and make it silghtly easier to do the rest of these tasks.

I couldn't use XCTUnwrap in TestUtils since I can't import XCTest because its in the TestTools project and not in a test target. So I decided to just return an optional and have the caller unwrap it.

Reviewed By: wx0165927473, samodom

Differential Revision: D35320238

fbshipit-source-id: 53c5f1a96b32e71eb8efc48a5a24a7920d2487a4