Fix improper indication of success when attempting to show a message dialog
Summary: In the class MessageDialog, the original Objective-C exited with a true value when show() was called but a bridge API request could not be created.  This was no doubt the wrong behavior and this diff updates it to return false as an early exit.  The type was originally converted to Swift maintaining this behavior.

Reviewed By: josueBrizuelaXD

Differential Revision: D36215258

fbshipit-source-id: 0cd0211054d05a4e20040cde86546babbb764d17