Fix a string comparison bug affecting AEM
Summary:
This has not been discovered by our unit tests because we use NSUserActivityTypeBrowsingWeb and it is a reference type (NSString *).

This should not have happened and I should have known better not to compare two NSString types in this manner.

Since we currently instruct advertisers to use application:open:url instead so application:continueUserActivity, then it is not being used.

This is now backed up with a new unit test checking for a strict raw value.

Reviewed By: samodom

Differential Revision: D36703886

fbshipit-source-id: 3dba28c5e6b3982428a8b67efa7f254ffab0a524