Adding profile for use in qa builds
Summary:
This is the profile that's currently used by jackalope builds that are consumed by QA.

The assumption is that if we land this, then we can build on machost and the matching cert will just be there.

Then we don't have to mess with autograph, it will just let us archive and export the archive as an ipa without additional steps.

If this works then the next step is to rework the sandcastle job in www to fetch this cert from the fbobjc/Tools directory so that we don't have to worry about expiration.

Test Plan: Land this then try in sandcastle.

Reviewers: josuebrizuelaxp, #sdk

Reviewed By: josuebrizuelaxp

Differential Revision: https://phabricator.intern.facebook.com/D36810846

Tags: publish_when_ready