Adding codesigning support for CoffeeShop
Summary:
Does a couple things.

1. Updates the generated coffeeshop target to use correct casing.
2. Adds files that autograph and the www build job will need in order to sign the files.
3. Includes a script to move the new files into the build.

Future optimizations would include sharing the script that moves these files but for now the duplication is not an issue and it will be easier to get both jobs working then refactor them to share code.

Test Plan:
N/A

This is all part of moving files to this repo and making sure that we can build artifacts once it becomes the source of truth. Currently the working code that people and sandcastle jobs rely on is located in fbsource/fbobjc/ios-sdk.

Changes to this repo effectively don't matter until we move everything over later.

NOTE: The hackbook build IPA job is known broken. Not a big deal right now. It needed to break in order to get the CI version working. I'll fix it later before we actually move anything.

Reviewers: danielhuri, ldelgado, jialichen, samodom, #sdk

Reviewed By: samodom

Differential Revision: https://phabricator.intern.facebook.com/D37099126

Tags: publish_when_ready