Removes unused entitlement file
Summary:
This is confusing because this file is never actually used. For signing in the nightly build we use the file `BUCK_code_sign_entitlements.plist`.

This is additionally confusing because we do not use BUCK but we need to call the file this because of what the code signing tool Autograph looks for.

Test Plan: N/A

Reviewers: danielhuri, yiweivivi, samodom, #sdk

Reviewed By: samodom

Differential Revision: https://phabricator.intern.facebook.com/D37468885

Tags: publish_when_ready