[Meta SDK] Implement KeychainStore to save userSession data
Summary:
The keychain is an encrypted container which will be used to store UserSession  (including user and authentication data).

The ```KeychainStore``` is a wrapper to manage official keychain APIs. It will be injected in ```LocalStorage```  as a dependency to store/access UserSession in keychain (The related diff is on the way).

Test Plan: The test will be covered in the LocalStorage diff

Reviewers: joesusnick, jwujoyce, pansy0419, #horizon_identity

Reviewed By: joesusnick

Differential Revision: https://phabricator.intern.facebook.com/D37986505

Tasks: T126405726