[Meta SDK] Remove IncomingURLHandling Protocol
Summary:
`IncomingURLHandling` is a protocol originally created to parse the incoming authentication URL to create a user session. That functionality is handled in `LoginResponse` and the protocol doesn't reflect the current design.

We want to remove the protocol since it is currently not being used and will not be used in the future.

Test Plan: Project should still build and pass all unit tests with deleted file.

Reviewers: joesusnick, pansy0419, xinh, #horizon_identity

Reviewed By: pansy0419

Differential Revision: https://phabricator.intern.facebook.com/D38369428