[iOS SDK] Fix bug in Hackbook CI Job
Summary:
Hackbook nightly job failed the other night. First thought was 'how is this possible? We have a CI test that mimics that build.'.

So I check the CI job on the diff that broke it and it looks like it failed silently. https://www.internalfb.com/intern/sandcastle/job/9007199991927573/insights

```
Exit code: 0

Output:
/bin/bash: line 2: .generate-projects.sh: command not found

// keeps doing stuff. What???
```

It's crazy that this failed silently. `command not found:` is not a 0 exit code!!!

I tried to fix it while introducing a different error and it passed again!
```
Output:
[1;33mWARNING: Expected xcodegen version is 2.25.0. You have 2.29.0.[0m
```

So I added the `|| exit 1;` which seems like it shouldn't have been necessary but did cause the job to fail which is good.

Also would have expected the "required" upload artifacts step to fail since there should not be artifacts at that path but somehow that was also succeeding.

Test Plan: Changed expected hackbook version in generate scripts so that the step would fail.

Reviewers: samodom, #sdk

Reviewed By: samodom

Differential Revision: https://phabricator.intern.facebook.com/D38459004

Tags: publish_when_ready