[Meta SDK] Introduce DependentAsActorInstance protocol
Summary: This protocol is a simplification of the other similar protocols but does not have resetting and does not have default implementations because actors cannot inherit anything.  This means that implementers must the boilerplate implementations as specified in the comments of the file.

Test Plan: sandcastle

Reviewers: joesusnick, adamcmiel, josuebrizuelaxp, #sdk, #horizon_identity

Reviewed By: josuebrizuelaxp

Differential Revision: https://phabricator.intern.facebook.com/D38626666

Tags: iOS_SDK_DI, swift-adoption