Deleted `FBSDKRestrictiveData.h/m`
Summary: We are not using this type anywhere other than tests, and it is internal, so I believe we should delete it instead of converting it to Swift. Let me know if you find evidence on why this should not be deleted.

Test Plan: sandcastle

Reviewers: samodom, #sdk

Reviewed By: samodom

Differential Revision: https://phabricator.intern.facebook.com/D38993832

Tasks: T130275998

Tags: opensource-needs-export, swift-adoption, accept2ship, publish_when_ready