Convert FBSDKServerConfigurationProvider.h/m to Swift
Summary:
Some context on why this type exists. Long ago there was a lot of over-exposure of types between kits using an internal umbrella header. Untangling this was complex and some types did in fact need some of their information exposed.

Effectively this type was created to avoid exposing all of `FBSDKServerConfigurationManager` to Login Kit. The irony is that to convert this pass-through type to Swift, we need to temporarily make `FBSDKServerConfigurationManager` public.

It still makes sense to me to do this since once `FBSDKServerConfigurationManager` is converted to Swift, we can make it private again and continue to expose the more limited interface through this type.

Other semi-interesting things about this changeset:

1. The nullability on the `loggingToken` property was incorrect. This fixes that.
1. LoginToolTip requires there to be text. Since text coming back is nullable, this won’t work. The only actual call site for this doesn’t actually check or use the returned error from the completion but adding a reasonable error still feels like the right thing to do here. So I introduced a new error type: `LoginTooltipError` with a single case: `missingTooltipText`. Alternately we could just return nil for the tooltip and everything will work the same as it does now.
1. Renames `ServerConfigurationManager` to `_ServerConfigurationManager` to keep it out of typeahead.
1. `DeviceLoginViewController` was using this pass-through type instead of using `ServerConfigurationManager` directly. To use `ServerConfigurationProvider` here we'd have to expose it to ObjC and subclass NSObject which seems not worth it since this kit is likely to be deprecated and deleted in the near future.

Test Plan: Build for regular and tv. This type does not have good unit tests and it is unclear what it would take to add them at this point since the type is pass-through to a singleton which makes it hard to test.

Reviewers: samodom, #sdk

Reviewed By: samodom

Differential Revision: https://phabricator.intern.facebook.com/D39057645

Tags: opensource-needs-export, swift-adoption, accept2ship, publish_when_ready