[iOS][HB] Update Hackbook Share Dialog test case PageID
Summary:
Previously the pageID that we used for testing is disabled so it is not showing up correctly in the composer and cause the Native Feed composer crash when launching. T126739139

In this diff, i'm updating the pageID to "104048449631599" (Menlo Park).

Test Plan:
| Browser Feed Composer | Native Feed Composer |
| can load the location correctly | no longer crashing, Next step is passing to Composer team to investigate why the location tag is not showing up.|
|{F765306029} | https://pxl.cl/2c5BM |

Reviewers: mshi, samodom, #sdk, #creation_growth_ios

Reviewed By: mshi

Differential Revision: https://phabricator.intern.facebook.com/D39116169

Tasks: T126739139

Tags: oncall, ddr_ml_flagged_advice, oncallteam-creation-growth