Convert FBSDKEventDeactivationManager.h/m to Swift
Summary:
Things to note:

Previously a crash in this type will show up looking like:
```
-[FBSDKEventDeactivationManager enable]
```

If I raise an NSException from Swift it looks like:
```
$s12FBSDKCoreKit24EventDeactivationManagerC6enableyyF + 234
```

With a sigtrap (ex: fatalError, out of index on array) it won't be handled at all. So effectively a Swift crash will not be captured by the crash handler.

I researched capturing sigtraps and found this quote from an Apple engineer, "Taking a step back, you seem to be implementing your own crash reporter. This is impossible to do correctly."
Source: https://developer.apple.com/forums/thread/113742
https://developer.apple.com/forums/thread/132556

Assuming it's an NSException from Swift, it is unclear whether this will be attributed correctly because of the logic in how we handle the method map.

Previously it would show up as `FBSDKEventDeactivationManager`. Now it's `FBSDKCoreKit.EventDeactivationManager`.

But we won't even try to string match on these names unless a memory address matches in the method map. Which brings us to the issue with a race condition around the method map. We create the method map on a background thread so there is a race condition that depends on when the crash takes place. If a crash happens before the method map is generated, then the crash is saved but the method map is not. So when the app opens again, the crash cannot be mapped to a method and the shield does nothing.

That's all aside from the fact that the
The example in the test plan does not seem to actually work when I run it locally. https://www.internalfb.com/intern/wiki/App-signals/Product_Test_Instructions/Crash_Shield_QA_Test_Plan/

Reviewed By: josueBrizuelaXD

Differential Revision: D39357303

fbshipit-source-id: 1567a778350f7893f4e638587962fcc35820b58e