Fix the new "Run script build phase" warning for Swiftlint in Xcode 14
Summary:
More context here:
https://github.com/realm/SwiftLint/issues/4015

The full warning is:
```
warning build: Run script build phase 'Run Swiftlint' will be run during every build because it does not specify any outputs. To address this warning, either add output dependencies to the script phase, or configure it to run in every build by unchecking "Based on dependency analysis" in the script phase.
```

The fix is to uncheck "Based on dependency analysis" for the run script build phase.

This adds a "alwaysOutOfDate = 1;" setting to the Xcode pbxproj file

This setting is turned on by using "basedOnDependencyAnalysis: false" in xcodegen

Reviewed By: samodom

Differential Revision: D39678897

fbshipit-source-id: 90ed96cab52c3e2a0be3082dc94884a773faa1c6