Eliminate use of FBSDKTestCoder from ServerConfigurationTests.swift
Summary:
Also fixed warnings:

```
Attempted to decode a collection type 'NSDictionary' (subclass of 'NSObject') for key 'NS.objects'. 'NSDictionary' requires its subclasses to be explicitly added to the allowed classes list but it is not present.
Attempted to decode a collection type 'NSString' (subclass of 'NSObject') for key 'NS.keys'. 'NSString' requires its subclasses to be explicitly added to the allowed classes list but it is not present.
```

Reviewed By: joesus

Differential Revision: D39778499

fbshipit-source-id: d039ae2ee6bb75557eb4926ba8a6fa4c5635d1da